refactor(pipeline): extract UI stages — __init__.py under 620 lines
All 5 UI pipeline stages (Piper + Glazier) moved to stages/ui.py:
- analyze_ui, define_ui_criteria, write_ui_tests, implement_ui, verify_ui

Static delegation stubs permanently anchored on RuntimePipeline for
methods that extracted stage handlers call via pipeline._*:
_extract_files_from_issue_body, _count_passing, _count_failing,
_parse_violation_files. These kept getting swept out during block
extractions — now placed before the UI delegations block.

__init__.py: 614 lines. No file in the package exceeds 600 LOC
except decompose.py (562) and __init__.py (614).

760 tests pass, 0 failures.
